+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 mobile_menu Changelog
 =====================
 
+Version 2.3.0 stable (2015-12-21)
+---------------------------------
+
+### Improved
+- Added the option to disable menu shadow (see #6)
+- The collapsible navigation feature is now optional (see #5)
+- Use the "vh" units for full height elements (see #4)
+
+### Fixed
+- Hide the mobile menu trigger initially via CSS (see #8)
+
+
 Version 2.2.4 stable (2015-11-19)
 ---------------------------------